Criando classe pedido - backend

+++ backend/src/main/java/br/com/fernandinesuplementos/loja/DTOs/ProductDto.java
@@ -1,14 +1,11 @@
 package br.com.fernandinesuplementos.loja.DTOs;
 
-import br.com.fernandinesuplementos.loja.entities.Category;
-import br.com.fernandinesuplementos.loja.entities.Subcategory;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -20,7 +17,7 @@ public class ProductDto implements Serializable {
     private String sku;
     private String name;
     private String description;
-    private BigDecimal unitPrice;
+    private Double unitPrice;
     private Long EAN;
     private String image;
     private boolean favorite;
